Finalisation Simulateur
- Passage des ID à Long (Permet de le set à null car c'est un Object et non un type primitif).

- Suppression des classes inutiles.

- Ajout de serializable partout.

- Modification du mouvement d'un camion, lorsqu'il est available il faut le rediriger vers la caserne, sinon vers l'incident.

+++ src/main/java/simulator/projet/repository/Irepository/ITruckRepository.java
@@ -1,6 +1,5 @@
 package simulator.projet.repository.Irepository;
 
-import org.springframework.stereotype.Service;
 import simulator.projet.model.Truck;
 import simulator.projet.model.viewModel.TruckViewModel;
 
@@ -11,10 +10,10 @@ public interface ITruckRepository {
 
     List<Truck> getTrucks();
 
-    Truck getTruck(int id);
+    Truck getTruck(Long id);
 
     Truck createTruck(TruckViewModel truck);
 
-    Truck updateTruck(int truck_id, int posX, int posY) ;
+    Truck updateTruck(Long truck_id, int posX, int posY);
 
-    }
+}
