+++ src/main/java/simulator/projet/repository/TruckRepository.java
@@ -1,7 +1,6 @@
 package simulator.projet.repository;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import simulator.projet.model.Truck;
 import simulator.projet.model.viewModel.TruckViewModel;
@@ -13,9 +12,8 @@ import java.util.List;
 public class TruckRepository extends BaseRepository<Truck> implements ITruckRepository {
 
     public TruckRepository(@Value("${api.baseUrl}") String baseUrl, @Value("${api.baseUrlManager}") String baseUrlManager) {
-        super(baseUrl,baseUrlManager);
+        super(baseUrl, baseUrlManager);
     }
-
 
 
     @Override
@@ -26,17 +24,17 @@ public class TruckRepository extends BaseRepository<Truck> implements ITruckRepo
 
     @Override
     public Truck createTruck(TruckViewModel truck) {
-        return this.postItem("/truck", truck, TruckViewModel.class , true);
+        return this.postItem("/truck", truck, TruckViewModel.class, true);
     }
 
     @Override
-    public Truck updateTruck(int truck_id, int posX, int posY ) {
-        return this.postItem("/truck/" + truck_id +"/posx/" + posX  + "/posy/" + posY,"" ,Truck.class , true);
+    public Truck updateTruck(Long truck_id, int posX, int posY) {
+        return this.postItem("/truck/" + truck_id + "/posx/" + posX + "/posy/" + posY, "", Truck.class, true);
     }
 
     @Override
-    public Truck getTruck(int id) {
-        return this.getItem("/truck/" + id, true) ;
+    public Truck getTruck(Long id) {
+        return this.getItem("/truck/" + id, true);
     }
